Added a release notes file for this release branch. Undertook system testing:
Usability: window is resizeable. confirmed readable by a non developer. Could do with tooltips
compliance: all files meet checkstyle. git commits are generally good
documentation: updated readme needed. javadoc is well used for the most part. gitk is understandable.
configuration: javafx runs on windows, although it cannot yet connect to the teaching server. this will be rectified when we switch to a local db. As long as the database is setup correctly it's possible to run on another machine that can connect to the teaching server. it would be worth creating a db setup method to create the required entries before startup if they don't exist yet.